Allow invoking fsync
The rationale is that often it is desired that directory metadata
be sync'd so changes become visible in the filesystem, in
particular after creating or renaming a file [0]. This is
necessary for some filesystems like ext4. To avoid races, the
call to fsync(2) must be passed the same fd that was used as the
target directory in the original operation (e. g. renamat(),
openat()).

In order for fsync to be effective, the fd must be equipped with
a full set of file operations kernel-side. Opening the directory
with `O_PATH` prevents this and attempts to fsync will fail with
EBADF. Thus an open variant `Dir::open_full()` is added that
omits the flag.

[0] https://lwn.net/Articles/457667/ (section: *When should you
    fsync?*)

Signed-off-by: Philipp Gesang <phg@phi-gamma.net>