Add/Refactor Dir::list() -> DirIter ctor
dir.list() consumes the dir and creates a DirIter. The underlying 'dup()' operation now
becomes explicit, one may need to call try_clone() first.

the list_self() and list_dir() using this now but are merely convinience wrappers and may (or
may not) deprecated in future.

Rationale: this simplifies the code a bit (list::open_dir() removed) and is closer to the low
level semantics.