start abstracting OS capabilities into features
Instead compiling conditionally on the target_os this defines features for every aspect that
can be different. Tested only on linux so far. Eventually the presence/absence of these
features should be determined by a build.rs script.

Meanwhile there are OS specific lists (only linux so far, please contribute more) which can be
activated when building with '--feature osname'.