Remove the '*_lite()' versions of open() and sub_dir()
With the flags builder the 'open_lite()' I introduced earlier can be removed. 'Dir::open(),
sub_dir()' defaults to include O_PATH when available as before. The flags builder does NOT include
O_PATH. 'Dir::list_self()' and 'Dir::list_dir()' already clone-upgrade the handle that removes
the O_PATH flag and thus will be not affected.

'Dir::list()' may fail when called with a Dir that is opened with O_PATH and the OS (linux)
does not support listing.